Send Auth request to SAP ID service.
We are able to get the metadata from SCI on its
metadata URL. We are able to parse it and get the
login URL from it.

We are then able to send a SAML auth request to it
using a browser redirect.

However, since we mutated the entity ID from what SCI
expects we get a error page on SCI.