Add token with user as authentication header
Use Joken/Jose libs to create JWT tokens. The token contains
the logged in user's information (like email,) and the token
is signed using a shared secret.

This token appears as "Bearer <token>" in the Authentication
header in forwarded http requests to the next service. That service
(either UI or API) should validate the token with the same secret and
accept the user's information contained in the token.

The elixir module LoginProxy.Jwt contains the functions to create and
verify the token.