Defer setting session cookie until redirect is followed
To support multi-tenancy via sub-domains, we have to support
sessions on multiple hosts. But our callback from SAP Cloud ID
always sends the saml assertion to one host that we have configured
on SCI. This causes the session to be set on the configured host,
not the one for a tenant other than the default.

To solve this, we defer setting a session cookie after SCI auth
until the following redirect has been followed. At that time, the
host is the required one. To do this, we attach a query param
to the original URL the user requested as the redirect address.
The query param is then used to
get the session id and the cookie is then set. We don't expose the
session id in the query param directly instead using a relay state.