Go back to setting session cookie on SAML assertion
For a while we had code to defer setting the session cookie so that
we could reuse the SP endpoint for multiple tenants. Only after the
redirect could we set the cookie. This is no longer needed since
we can create one SP endpoint per tenant.

It also removes the extra RelayState that we set as query params
in the redirect. Those params were causing kora-ui to not go to the
default URL. So it fixes a minor issue seen on first log in.